Stage 4 layouts+modifiers, text/font primitives, demo reel showcase
Continues the MoonLight migration: adds the modifier and layout libraries (6 modifiers, 12 layouts), a bitmap-font text renderer with a TextEffect, and grows DemoReel into a hands-free showcase that cycles every effect with a random palette and an on-screen name label. Also folds in the CodeRabbit review of the previous effect PR — 14 effect fixes, several of which are deliberate improvements over MoonLight, registered in a new log. Ant and PacMan (1D) are removed; a real 2D/3D PacMan is backlogged.

KPI: tick:3536us(FPS:282)

Core
- math8: map8 now reaches its range top (lo + in*(hi-lo)/255, FastLED's documented map8), so a one-step span like a 1-row audio bar is reachable instead of collapsing to 0.

Light domain
- draw: added the glyph/text bitmap-font blitter (draw::glyph reads columns MSB-first — leftmost pixel is bit 7 — so text isn't left-right mirrored; draw::text lays out multi-line strings, newline wraps down a font height).
- fonts: two flash-resident console fonts (4x6, 6x8) carried from the public raster-fonts set; more are pure data to add later.
- TextEffect: renders a multi-line string (the shared text-area control) in a selectable font — static top-left by default, an optional horizontal marquee, palette-coloured; default text "projectMM".
- Modifiers (6, MoonLight): Mirror, Transpose, Circle, Block, Pinwheel, RippleXZ — coordinate transforms on ModifierBase (modifyLogicalSize + modifyLogical), the 1D→2D/3D expanders (Circle/Block) among them.
- Layouts (12, MoonLight): SingleRow, SingleColumn, Panel, Panels, Cube, Human Sized Cube, Tubes, Ring, Rings 241, Spiral, Car Lights, Toronto Bar Gourds — emit their LED positions through LayoutBase (lightCount + forEachCoord).
- DemoReelEffect: added randomPalette (a fresh palette each cycle) and showName (overlay the playing effect's name, top-left) — both default on, making the reel a showcase/test tool for the whole library.
- MoonLiveEffect: default script is now random pixels (an always-visible starting example) instead of solid blue.
- Effect fixes (CodeRabbit): FreqSaws integrates each band once per frame (grid-width-independent speed); SphereMove uses float time (smooth motion); Lissajous maps a size-1 axis to 0 (draws on thin grids); PaintBrush spans the full grid on large panels; FixedRectangle clears the W channel on coloured RGBW tiles; GameOfLife guards the 3D neighbour count against the 9-entry B/S tables and gives injected/mutated cells a non-zero colour; GEQ/StarSky widen index/peak types for large installations; Tetrix uses wrap-safe deadlines; RubiksCube keeps its projection (perf note in code).

UI
- app.js: (from the prior batch) the module picker sorts alphabetically — the 18 new modules slot in by name.

Scripts / MoonDeck
- moondeck.py: the raw-<img> card renderer preserves the alt attribute (markdownlint + accessibility).

Tests
- unit_draw: pin the glyph orientation (the 'L' probe — bar on the left, foot at the bottom — catches a mirror/flip regression), draw::line shorten rounding, and draw::blur vs a reference.
- unit_TextEffect: static render, newline wrap, and scroll safety at any grid size.
- unit_math8: map8 reaches both range ends including one-step spans.
- unit_DemoReelEffect: clears the buffer between hosted renders so the render check measures each host.
- unit_GameOfLifeEffect: a dense 3D neighbourhood steps without an out-of-bounds B/S read.

Docs / CI
- effects.md / modifiers.md / layouts.md: blocks for the new modules (alphabetical within each library section); WLED-origin effect previews cross-linked from WLED-Utils (scottrbailey) with credit; StarSky / Car Lights / Toronto Bar Gourds previews moved from the image pool into assets and wired.
- moonlight-improvements.md (new): registers each deliberate divergence from MoonLight that improves an effect (map8 bar heights, FreqSaws timing, SphereMove smoothness, Lissajous thin-grid, PaintBrush coverage, FixedRectangle RGBW), so improvements are decisions on record.
- backlog-light.md: a real 2D/3D PacMan, and a fix for DemoReel extruding hosted 1D/2D effects (the first attempt crashed and was reverted).
- moonlight_images: drained the layouts folder — controls-panel screenshots deleted, render gifs moved to assets.

Reviews
- 🐇 map8 upper bound unreachable — fixed (FastLED-documented form) + test; logged as an improvement.
- 🐇 FreqSaws per-column band physics; SphereMove integer time; Lissajous size-1; PaintBrush uint8 bounds + off-by-one; FixedRectangle RGBW white; GameOfLife 3D OOB + zero-colour marker; GEQ/StarSky index widths; Tetrix wrap — all fixed; visible ones registered in moonlight-improvements.md.
- 🐇 DemoReel extrude hosted lower-D effects — attempted, crashed the suite (heap/vtable smash), reverted + backlogged with root-cause direction.
- 🐇 moondeck raw-img alt text; effects.md driver link → concrete page — fixed.
- 🐇 GEQ3D cols/NUM_BANDS narrow-grid + frame-throttled sweep — accepted as faithful to MoonLight; in moonlight-fidelity-tensions.md.
- 🐇 HueDriver persist-by-ID — deferred (control-mechanism redesign, not a bug).

+++ src/light/effects/SphereMoveEffect.h
@@ -22,10 +22,10 @@ namespace mm {
 //
 // Prior art: MoonLight's SphereMove (E_MoonModules / MoonModules). The origin oscillator math
 // (millis()/(100-speed)/6.4, the sin/cos origin path), the diameter = 2 + sin(ti/3) breathing,
-// the one-unit shell test, and the millis()/50 + random8(64) palette index are reproduced exactly,
-// written fresh on EffectBase + the shared draw primitives. MoonLight does the first divide
-// (millis()/(100-speed)) as integer division, then /6.4 as float — reproduced here so time_interval
-// matches bit-for-bit, not all-float.
+// the one-unit shell test, and the millis()/50 + random8(64) palette index are reproduced.
+// time_interval is computed in full float (ms and 100-speed as float through the whole
+// expression) so the origin sweep and diameter breathing integrate continuously rather than in
+// quantised integer-time steps — a smooth sweep at every speed.
 class SphereMoveEffect : public EffectBase {
 public:
     const char* tags() const override { return "💫🧊"; }  // MoonLight origin · 3D-native
@@ -51,11 +51,11 @@ public:
 
         const uint32_t ms = elapsed();
 
-        // Origin oscillator. MoonLight: millis()/(100-speed)/6.4, where the FIRST divide is integer
-        // (millis() and speed are integers) and only /6.4 is float — reproduced exactly. The divisor
-        // 100-speed (speed clamped 0..99 → divisor 1..100, never 0) and the (256-128)/20.0 == 6.4
-        // factor are MoonLight's exact constants.
-        const float time_interval = static_cast<float>(ms / static_cast<uint32_t>(100 - speed)) / 6.4f;
+        // Origin oscillator: time_interval = ms / (100-speed) / 6.4, all-float so time advances
+        // continuously (integer-dividing ms/(100-speed) first would truncate the sub-step time and
+        // quantise the sweep). The divisor 100-speed (speed clamped 0..99 → divisor 1..100, never 0)
+        // and the (256-128)/20.0 == 6.4 factor are MoonLight's exact constants.
+        const float time_interval = static_cast<float>(ms) / static_cast<float>(100 - speed) / 6.4f;
 
         const float ox = w / 2.0f * (1.0f + sinf(time_interval));
         const float oy = h / 2.0f * (1.0f + cosf(time_interval));
